api changes

+++ app/Models/User.php
@@ -64,7 +64,7 @@ class User extends Authenticatable
      */
     public function setPhoneAttribute($value)
     {
-        $this->attributes['phone'] = (int)str_replace("-", "", $value);
+        $this->attributes['phone'] = str_replace("-", "", $value);
     }
     /**
      * Get the user's Date Of Birth.
